requirements.txt
altair==4.2.0
attrs==22.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.0.4
colorama==0.4.6
covid==2.5.2
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.14
daterangestr==0.0.3
DateTime==5.0
decorator==5.1.1
distro==1.8.0
entrypoints==0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Simple-GeoIP==0.2.4
fonttools==4.38.0
gcloud==0.18.3
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.30
googleapis-common-protos==1.58.0
googletrans==3.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipaddress==1.0.23
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jws==0.1.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
mysql-connector-python==8.0.32
newsapi-python==0.2.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.21.6
oauth2client==4.1.3
openweather==0.11
packaging==23.0
pandas==1.1.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
Pyrebase==3.0.27
Pyrebase4==4.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.2.0
rsa==4.9
scikit-learn==0.21.3
scipy==1.7.3
semver==2.13.0
setuptools==66.0.0
sgmllib3k==1.0.0
simple-geoip==0.1.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
streamlit==1.9.0
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.21.1
vaderSentiment==3.3.2
validators==0.20.0
watchdog==2.2.1
Werkzeug==2.2.2
zipp==3.11.0
zope.interface==5.5.2